[#26] - feat 인디게이터 코드 분할
